+++ app/src/main/java/com/example/pillmate/domain/usecase/LogTaskUseCase.kt
@@ -45,7 +45,7 @@ class LogTaskUseCase(
                 medicationRepository.logInventoryChange(
                     profileId = profileId,
                     medicationId = sourceId,
-                    amount = -dose.toInt(),
+                    amount = -dose,
                     reason = "Taken"
                 )
             }
